Put in time zoning and query execution at departure time
The URL executed on the Google API is given the UTC time since epoch,
which is populated with time zone from the machine running the code.
Queries should be given specific time zones to avoid confusion (they
are converted to the computer’s time zone, which then get’s converted
to UTC since epoch).